portmapper: don't compile linux-only code on Windows
